Update core-geometric-system.mjs
I messed it up earlier. Sorry about that. I'm working on it. The CgsSphericalCap class has been deprecated. The volume of a spherical cap is a property of the CgsSphere class. It can be calculated two different ways. I was trying to combine them and ended up somewhere in between. This should work.